Modify handlewidget as baseclass of buttonwidgets
